pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentiment-dashboard"
version = "0.1.0"
description = "Internal BlazeTV host popularity dashboard monorepo"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["apps*", "jobs*", "packages*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests", "apps/api/tests"]

